Output and Saving Controls

These callbacks extend the output and saving controls available during time stepping.

DiffEqCallbacks.SavingCallback — Function

SavingCallback(save_func, saved_values::SavedValues;
    saveat = Vector{eltype(saved_values.t)}(),
    save_everystep = isempty(saveat),
    save_start = true,
    tdir = 1)

The saving callback lets you define a function save_func(u, t, integrator) which returns quantities of interest that shall be saved.

Arguments

save_func(u, t, integrator) returns the quantities which shall be saved. Note that this should allocate the output (not as a view to u).

saved_values::SavedValues is the types that save_func will return, i.e. save_func(t, u, integrator)::savevalType. It's specified via SavedValues(typeof(t),savevalType), i.e. give the type for time and the type that save_func will output (or higher compatible type).

Keyword Arguments

saveat mimics saveat in solve from solve.

save_everystep mimics save_everystep from solve.

save_start mimics save_start from solve.

save_end mimics save_end from solve.

tdir should be sign(tspan[end]-tspan[1]). It defaults to 1 and should be adapted if tspan[1] > tspan[end].

The outputted values are saved into saved_values. Time points are found via saved_values.t and the values are saved_values.saveval.

DiffEqCallbacks.FunctionCallingCallback — Function

FunctionCallingCallback(func;
    funcat = Vector{Float64}(),
    func_everystep = isempty(funcat),
    func_start = true,
    tdir = 1)

The function calling callback lets you define a function func(u,t,integrator) which gets calls at the time points of interest. The constructor is:

func(u, t, integrator) is the function to be called.

funcat values or interval that the function is sure to be evaluated at.

func_everystep whether to call the function after each integrator step.

func_start whether the function is called the initial condition.

tdir should be sign(tspan[end]-tspan[1]). It defaults to 1 and should be adapted if tspan[1] > tspan[end].

Saving Example

In this example, we will solve a matrix equation and at each step save a tuple of values which contains the current trace and the norm of the matrix. We build the SavedValues cache to use Float64 for time and Tuple{Float64,Float64} for the saved values, and then call the solver with the callback.

using DiffEqCallbacks, OrdinaryDiffEq, LinearAlgebra
prob = ODEProblem((du, u, p, t) -> du .= u, rand(4, 4), (0.0, 1.0))
saved_values = SavedValues(Float64, Tuple{Float64, Float64})
cb = SavingCallback((u, t, integrator) -> (tr(u), norm(u)), saved_values)
sol = solve(prob, Tsit5(), callback = cb)

print(saved_values.saveval)

[(1.7699513442092645, 2.161297869631493), (1.9564853600158403, 2.389075640074483), (2.5082179306899737, 3.0627994876286366), (3.5082147109042534, 4.283901365816731), (4.811226364995878, 5.87501646697867)]

Note that the values are retrieved from the cache as .saveval, and the time points are found as .t. If we want to control the saved times, we use saveat in the callback. The save controls like saveat act analogously to how they act in the solve function.

saved_values = SavedValues(Float64, Tuple{Float64, Float64})
cb = SavingCallback((u, t, integrator) -> (tr(u), norm(u)), saved_values,
    saveat = 0.0:0.1:1.0)
sol = solve(prob, Tsit5(), callback = cb)
print(saved_values.saveval)
print(saved_values.t)

[(1.7699513442092645, 2.161297869631493), (1.9560987520702455, 2.3886035508659686), (2.161823531512397, 2.6398152742805303), (2.3891842660347082, 2.917446788145934), (2.640457607485812, 3.2242781252616317), (2.918156284289158, 3.563377592904171), (3.2250603606404384, 3.9381399436147664), (3.5642443049933137, 4.3523194286845), (3.9390994736040694, 4.810057252941318), (4.353376443050541, 5.315933267234844), (4.811226364995878, 5.87501646697867)][0.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]